+++ alchemic/ALCValue+Injection.m
@@ -231,7 +231,7 @@ scalarMethodArgumentInjector(CGRect, CGRect)
             case 1:
                 return values[0];
             default:
-                throwException(AlchemicIncorrectNumberOfValuesException, @"Expected 1, got %lu", values.count);
+                throwException(AlchemicIncorrectNumberOfValuesException, @"Expected 1, got %lu", (unsigned long) values.count);
         }
     }
